[#269] Instead of patching rubytree, creating a custom sub class and using that everywhere
